chore: Rename account and account code modules
