+++ bin/bench-tx/src/main.rs
@@ -5,7 +5,7 @@ use std::path::Path;
 
 use anyhow::Context;
 use miden_lib::note::create_p2id_note;
-use miden_lib::testing::account::MockAccountExt;
+use miden_lib::testing::mock_account::MockAccountExt;
 use miden_lib::testing::account_component::IncrNonceAuthComponent;
 use miden_objects::account::{Account, AccountId, AccountStorageMode, AccountType};
 use miden_objects::asset::{Asset, FungibleAsset};
