+++ crates/miden-testing/src/kernel_tests/tx/test_account_interface.rs
@@ -2,7 +2,7 @@ use alloc::vec::Vec;
 
 use assert_matches::assert_matches;
 use miden_lib::note::{create_p2id_note, create_p2ide_note};
-use miden_lib::testing::account::MockAccountExt;
+use miden_lib::testing::mock_account::MockAccountExt;
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::account::{Account, AccountId};
 use miden_objects::asset::FungibleAsset;
